Rename descriptors/calc.py to descriptors/quippy.py
